feat: split sector tracker
