+++ damocles-manager/cmd/damocles-manager/internal/util_sealer_proving.go
@@ -29,7 +29,6 @@ import (
 	"github.com/filecoin-project/venus/venus-shared/actors/builtin/miner"
 	"github.com/filecoin-project/venus/venus-shared/types"
 
-	ffi "github.com/filecoin-project/filecoin-ffi"
 	"github.com/ipfs-force-community/damocles/damocles-manager/core"
 	"github.com/ipfs-force-community/damocles/damocles-manager/modules/policy"
 	"github.com/ipfs-force-community/damocles/damocles-manager/modules/util"
@@ -830,7 +829,11 @@ var utilSealerProvingWinningVanillaCmd = &cli.Command{
 		},
 	},
 	Action: func(cctx *cli.Context) error {
-		api, actx, astop, err := extractAPI(cctx)
+		var (
+			prover   core.Prover
+			verifier core.Verifier
+		)
+		api, actx, astop, err := extractAPI(cctx, &prover, &verifier)
 		if err != nil {
 			return err
 		}
@@ -878,7 +881,7 @@ var utilSealerProvingWinningVanillaCmd = &cli.Command{
 
 		randomness := make(abi.PoStRandomness, abi.RandomnessLength)
 
-		challenges, err := ffi.GeneratePoStFallbackSectorChallenges(abi.RegisteredPoStProof_StackedDrgWinning32GiBV1, sectorID.Miner, randomness, []abi.SectorNumber{sectorID.Number})
+		challenges, err := prover.GeneratePoStFallbackSectorChallenges(actx, abi.RegisteredPoStProof_StackedDrgWinning32GiBV1, sectorID.Miner, randomness, []abi.SectorNumber{sectorID.Number})
 		if err != nil {
 			return fmt.Errorf("generate challenge for sector %s: %w", sealedFileName, err)
 		}
@@ -890,7 +893,7 @@ var utilSealerProvingWinningVanillaCmd = &cli.Command{
 
 		slog.Infof("%d challenge generated", len(challenge))
 
-		vannilla, err := ffi.GenerateSingleVanillaProof(core.FFIPrivateSectorInfo{
+		vannilla, err := prover.GenerateSingleVanillaProof(actx, core.FFIPrivateSectorInfo{
 			SectorInfo:       sectorInfo,
 			PoStProofType:    abi.RegisteredPoStProof_StackedDrgWinning32GiBV1,
 			CacheDirPath:     cacheDirPath,
@@ -902,15 +905,14 @@ var utilSealerProvingWinningVanillaCmd = &cli.Command{
 
 		slog.Infof("vannilla generated with %d bytes", len(vannilla))
 
-		proofs, err := ffi.GenerateWinningPoStWithVanilla(abi.RegisteredPoStProof_StackedDrgWinning32GiBV1, sectorID.Miner, randomness, [][]byte{vannilla})
+		proofs, err := prover.GenerateWinningPoStWithVanilla(actx, abi.RegisteredPoStProof_StackedDrgWinning32GiBV1, sectorID.Miner, randomness, [][]byte{vannilla})
 		if err != nil {
 			return fmt.Errorf("generate winning post with vannilla for %s: %w", sealedFileName, err)
 		}
 
 		slog.Infof("proof generated with %d bytes", len(proofs[0].ProofBytes))
 
-		randomness[31] &= 0x3f
-		verified, err := ffi.VerifyWinningPoSt(core.WinningPoStVerifyInfo{
+		verified, err := verifier.VerifyWinningPoSt(actx, core.WinningPoStVerifyInfo{
 			Randomness:        randomness,
 			Proofs:            proofs,
 			ChallengedSectors: []core.SectorInfo{sectorInfo},
